Creates a table with events from the image level annotation information.

Creates a table with events from the image level annotation information.

watchme_aggregate(df, min_no_pictures = 1)

Arguments

df

the data created by using watchme_prepare_data

min_no_pictures

the minimal number of images for defining an event. Default is 1.

Value

A tbl_df with

event code

start time (POSIXt)

end time (POSIXt)

no of pictures in the event

index of the first picture in the event

index of the last picture in the event

group of the code

meaning of the code

duration of the event in seconds

the attribute dico tibble for defining the codes with at least Code and Meaning column, possibly Group column for having groups of codes (e.g. sport encompasses running and swimming)

event index, , , and event_code (character).

Examples

data('coding_example')
event_table <- watchme_aggregate(df = coding_example)
event_table

#> # A tibble: 66 × 9
#>    event_code          start_time            end_time no_pictures start_picture
#>         <chr>              <dttm>              <dttm>       <int>         <int>
#> 1         01. 2015-06-03 08:45:09 2015-06-03 08:45:09           1             7
#> 2         01. 2015-06-03 09:13:34 2015-06-03 09:13:34           1            50
#> 3         01. 2015-06-03 09:15:24 2015-06-03 09:15:24           1            53
#> 4         01. 2015-06-03 09:16:46 2015-06-03 09:22:23           8            55
#> 5         01. 2015-06-03 09:37:20 2015-06-03 09:37:53           2            81
#> 6         01. 2015-06-03 10:48:40 2015-06-03 10:48:40           1           180
#> 7         01. 2015-06-03 13:21:36 2015-06-03 13:22:43           3           309
#> 8         01. 2015-06-03 15:06:04 2015-06-03 15:06:47           2           449
#> 9         01. 2015-06-03 15:10:01 2015-06-03 15:10:01           1           456
#> 10        01. 2015-06-04 06:13:30 2015-06-04 06:16:37           5           801
#> # ... with 56 more rows, and 4 more variables: end_picture <int>, group <chr>,
#> #   meaning <fctr>, duration <dbl>

event_table2 <- watchme_aggregate(df = coding_example, min_no_pictures = 2)
event_table2

#> # A tibble: 29 × 9
#>    event_code          start_time            end_time no_pictures start_picture
#>         <chr>              <dttm>              <dttm>       <int>         <int>
#> 1         01. 2015-06-03 09:16:46 2015-06-03 09:22:23           8            55
#> 2         01. 2015-06-03 09:37:20 2015-06-03 09:37:53           2            81
#> 3         01. 2015-06-03 13:21:36 2015-06-03 13:22:43           3           309
#> 4         01. 2015-06-03 15:06:04 2015-06-03 15:06:47           2           449
#> 5         01. 2015-06-04 06:13:30 2015-06-04 06:16:37           5           801
#> 6         01. 2015-06-04 06:21:13 2015-06-04 06:35:26          19           811
#> 7         01. 2015-06-04 06:42:46 2015-06-04 06:43:24           2           840
#> 8         01. 2015-06-04 07:01:41 2015-06-04 07:02:14           2           862
#> 9         01. 2015-06-04 07:03:37 2015-06-04 07:04:51           3           865
#> 10        01. 2015-06-04 07:06:09 2015-06-04 07:09:14           6           869
#> # ... with 19 more rows, and 4 more variables: end_picture <int>, group <chr>,
#> #   meaning <fctr>, duration <dbl>