Combine objects with same pictures but different codes

Combine objects with same pictures but different codes

watchme_combine_results(results_list, common_codes = c("non codable"))

Arguments

results_list

a list of tibble created by watchme_prepare_data.

common_codes

One or several codes that are common between the objects.

Value

A tibble with

participantID Name or ID number of the participant (character)

image_time Path or name of the image in order to be able to identify duplicates (character)

image_time Time and date of each image (POSIXt)

booleanCodes columns of boolean, indicating if a given code was given to a given picture. codes is a condensed form of this.

the attribute dico tibble for defining the codes with at least Code and Meaning column, possibly Group column for having groups of codes (e.g. sport encompasses running and swimming)

Details

The motivation was to create a single object based on several coding files in the CHAI project: each pass/group of codes such as cooking and travel resulted in one file, and we want to combine them.

Forcommon_codes codes, the annotation are merged: if an image has such a code in an object and not in the other, the code is assigned to the image in the resulting object. The code does not look for conflicts.

Examples

passes <- c("CK", "IO", "OP", "PM", "TP")

create_pass_results <- function(pass){
  path_results <- system.file('extdata', paste0("oneday_", pass, ".csv"),
  package = 'watchme')
  sep_results <- "\t"
  path_dico <-  system.file('extdata', paste0("dico_coding_2016_01_", pass, ".csv"),
   package = 'watchme')
  sep_dico <- ';'

  results <- watchme_prepare_data(path_results = path_results,
                                  sep_results = sep_results,
                                  path_dico = path_dico,
                                  sep_dico = sep_dico,
                                  tz = "Asia/Kolkata")
  results$image_path <- gsub('\"', "", results$image_path)
  results
}

results_list <- purrr::map(passes, create_pass_results)

#> Parsed with column specification:
#> cols(
#>   Filename = col_character(),
#>   `EXIF:Date Taken [Y-m-d_H-M-S]` = col_character(),
#>   `IPTC:Keywords` = col_character(),
#>   X4 = col_character()
#> )

#> Warning: hms, hm and ms usage is deprecated, please use HMS, HM or MS instead. Deprecated in version '1.5.6'.

#> Parsed with column specification:
#> cols(
#>   `"Filename` = col_character(),
#>   `EXIF:Date Taken [Y-m-d_H-M-S]` = col_character(),
#>   `IPTC:Keywords` = col_character(),
#>   `"` = col_character()
#> )

#> Warning: hms, hm and ms usage is deprecated, please use HMS, HM or MS instead. Deprecated in version '1.5.6'.

#> Parsed with column specification:
#> cols(
#>   Filename = col_character(),
#>   `EXIF:Date Taken [Y-m-d_H-M-S]` = col_character(),
#>   `IPTC:Keywords` = col_character(),
#>   X4 = col_character()
#> )

#> Warning: hms, hm and ms usage is deprecated, please use HMS, HM or MS instead. Deprecated in version '1.5.6'.

#> Parsed with column specification:
#> cols(
#>   Filename = col_character(),
#>   `EXIF:Date Taken [Y-m-d_H-M-S]` = col_character(),
#>   `IPTC:Keywords` = col_character(),
#>   X4 = col_character()
#> )

#> Warning: hms, hm and ms usage is deprecated, please use HMS, HM or MS instead. Deprecated in version '1.5.6'.

#> Parsed with column specification:
#> cols(
#>   Filename = col_character(),
#>   `EXIF:Date Taken [Y-m-d_H-M-S]` = col_character(),
#>   `IPTC:Keywords` = col_character(),
#>   X4 = col_character()
#> )

#> Warning: hms, hm and ms usage is deprecated, please use HMS, HM or MS instead. Deprecated in version '1.5.6'.

oneday_results <- watchme_combine_results(results_list,
common_codes = "non_codable")
oneday_results

#> # A tibble: 1,665 × 26
#>                           image_path          image_time participant_id
#>                                <chr>              <dttm>          <chr>
#> 1  B00000000_21I6LU_20150527_073822E 2015-05-27 07:38:01          no_id
#> 2  B00000001_21I6LU_20150527_073823E 2015-05-27 07:38:04          no_id
#> 3  B00000002_21I6LU_20150527_073824E 2015-05-27 07:38:06          no_id
#> 4  B00000003_21I6LU_20150527_073825E 2015-05-27 07:38:09          no_id
#> 5  B00000004_21I6LU_20150527_073826E 2015-05-27 07:38:12          no_id
#> 6  B00000005_21I6LU_20150527_073826E 2015-05-27 07:38:14          no_id
#> 7  B00000006_21I6LU_20150527_073826E 2015-05-27 07:38:17          no_id
#> 8  B00000007_21I6LU_20150527_073827E 2015-05-27 07:38:19          no_id
#> 9  B00000008_21I6LU_20150527_073828E 2015-05-27 07:38:22          no_id
#> 10 B00000009_21I6LU_20150527_080738E 2015-05-27 08:07:37          no_id
#> # ... with 1,655 more rows, and 23 more variables: biomass_cooking_unit <lgl>,
#> #   lpg_stove <lgl>, other_cooking_unit <lgl>, food_preparation <lgl>,
#> #   eating <lgl>, presence_in_the_kitchen <lgl>, indoors <lgl>, outdoors <lgl>,
#> #   in_vehicle <lgl>, mixed <lgl>, presence_at_office_or_shop <lgl>,
#> #   presence_at_work_field <lgl>, presence_in_industry <lgl>,
#> #   presence_in_informal_work <lgl>, diesel_generator <lgl>, smoking <lgl>,
#> #   visible_flame_or_smoke <lgl>, travel_by_bus <lgl>, travel_by_bicycle <lgl>,
#> #   travel_by_auto <lgl>, travel_by_motorcycle <lgl>,
#> #   participant_presence_on_road <lgl>, non_codable <lgl>