Calculates interrater agreement using the

Calculates interrater agreement using the irr package. The unit of comparison is one picture.

watchme_ira(results_list, names_list = NULL, one_to_one = TRUE,
  by_code = FALSE)

Arguments

results_list

a list of tibble created by watchme_prepare_data.

names_list

(optional) a vector of names for the coders. It must be the same length as results_list and contains different names.

one_to_one

a boolean indicating whether Cohen's kappa should be calculated for each possible pair of coders in case of more than 2 coders, instead of Fleiss's Kappa for all coders at the same time.

by_code

boolean indicating whether the IRR should be calculated for each code separately, or for annotations as a whole.

Value

A tbl_df presenting the results of a call to the irr function. If there are only two raters the called function is kappa2, unweighted. If there are more than two raters and one_to_one is FALSE, the called function is kappam.fleiss.

Examples

data('coding1')
data('coding2')
# With two coders
results_list <- list(coding1, coding2)
names_list <- c('Cain', 'Abel')
watchme_ira(results_list, names_list = names_list)

#> # A tibble: 1 × 7
#>                                             method pictures agreed_on
#>                                             <fctr>    <int>     <int>
#> 1 Cohen's Kappa for 2 Raters (Weights: unweighted)     1263      1230
#> # ... with 4 more variables: raters <fctr>, Kappa <dbl>, z <dbl>, p_value <dbl>

watchme_ira(results_list, names_list = names_list, by_code = TRUE)

#> # A tibble: 5 × 8
#>          code                                           method pictures
#>         <chr>                                           <fctr>    <int>
#> 1     indoors Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 2    outdoors Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 3  in_vehicle Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 4       mixed Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 5 non_codable Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> # ... with 5 more variables: agreed_on <int>, raters <fctr>, Kappa <dbl>,
#> #   z <dbl>, p_value <dbl>

# With three coders
results_list2 <- list(coding1, coding1, coding2)
names_list2 <- c('Riri', 'Fifi', 'Loulou')
watchme_ira(results_list2, names_list = names_list2, one_to_one = FALSE)

#> # A tibble: 1 × 7
#>                       method pictures agreed_on             raters     Kappa
#>                       <fctr>    <int>     <int>             <fctr>     <dbl>
#> 1 Fleiss' Kappa for m Raters     1263      1230 Riri, Fifi, Loulou 0.8964106
#> # ... with 2 more variables: z <dbl>, p_value <dbl>

watchme_ira(results_list2, names_list = names_list2, one_to_one = TRUE)

#> # A tibble: 3 × 7
#>                                             method pictures agreed_on
#>                                             <fctr>    <int>     <int>
#> 1 Cohen's Kappa for 2 Raters (Weights: unweighted)     1263      1263
#> 2 Cohen's Kappa for 2 Raters (Weights: unweighted)     1263      1230
#> 3 Cohen's Kappa for 2 Raters (Weights: unweighted)     1263      1230
#> # ... with 4 more variables: raters <chr>, Kappa <dbl>, z <dbl>, p_value <dbl>

watchme_ira(results_list2, names_list = names_list2, one_to_one = FALSE, by_code = TRUE)

#> # A tibble: 5 × 8
#>          code                     method pictures agreed_on             raters
#>         <chr>                     <fctr>    <int>     <int>             <fctr>
#> 1     indoors Fleiss' Kappa for m Raters     1263      1251 Riri, Fifi, Loulou
#> 2    outdoors Fleiss' Kappa for m Raters     1263      1232 Riri, Fifi, Loulou
#> 3  in_vehicle Fleiss' Kappa for m Raters     1263      1263 Riri, Fifi, Loulou
#> 4       mixed Fleiss' Kappa for m Raters     1263      1241 Riri, Fifi, Loulou
#> 5 non_codable Fleiss' Kappa for m Raters     1263      1263 Riri, Fifi, Loulou
#> # ... with 3 more variables: Kappa <dbl>, z <dbl>, p_value <dbl>

watchme_ira(results_list2, names_list = names_list2, one_to_one = TRUE, by_code = TRUE)

#> # A tibble: 15 × 8
#>           code                                           method pictures
#>          <chr>                                           <fctr>    <int>
#> 1      indoors Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 2      indoors Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 3      indoors Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 4     outdoors Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 5     outdoors Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 6     outdoors Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 7   in_vehicle Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 8   in_vehicle Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 9   in_vehicle Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 10       mixed Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 11       mixed Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 12       mixed Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 13 non_codable Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 14 non_codable Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> 15 non_codable Cohen's Kappa for 2 Raters (Weights: unweighted)     1263
#> # ... with 5 more variables: agreed_on <int>, raters <chr>, Kappa <dbl>,
#> #   z <dbl>, p_value <dbl>