Creates a wearableCamImages object from information read in a csv file.

Creates a wearableCamImages object from information read in a csv file.

watchme_prepare_data(path_results, sep_results, quote_sign = "'",
  participant_id = "no_id", path_dico, sep_dico, tz = "Asia/Kolkata",
  robust_reading = FALSE)

Arguments

path_results

the path to the file with coding results

sep_results

the separator in the file with coding results

quote_sign

the quote argument of read.table for the results, default is "\'"

participant_id

participant ID, if available.

path_dico

the path to the file with the list of annotations

sep_dico

the separator in the file with the list of annotations

tz

timezone

robust_reading

FALSE by default. When TRUE the annotation column is the whole line: the file is read once for finding photo paths and times, and once again for finding code.

Value

A tibble with

participant_id Name or ID number of the participant (character)

image_time Path or name of the image in order to be able to identify duplicates (character)

image_time Time and date of each image (POSIXt)

booleanCodes columns of boolean, indicating if a given code was given to a given picture. codes is a condensed form of this.

the attribute dico tibble for defining the codes with at least Code and Meaning column, possibly Group column for having groups of codes (e.g. sport encompasses running and swimming)

Details

Please check the format that both files should have by looking at the provided example data. However you could consider write your own function for converting the input instead of having to re-format all your existing data, which we could even add to the package.

Examples

path_results <- system.file('extdata', 'image_level_pinocchio.csv', package = 'watchme')
sep_results <- ','
path_dico <-  system.file('extdata', 'dicoCoding_pinocchio.csv', package = 'watchme')
sep_dico <- ';'
data_pictures <- watchme_prepare_data(path_results=path_results, sep_results=sep_results,
              path_dico=path_dico, sep_dico=sep_dico)

#> Parsed with column specification:
#> cols(
#>   name = col_character(),
#>   image_path = col_character(),
#>   image_time = col_datetime(format = ""),
#>   annotation = col_character()
#> )

#> Warning: hms, hm and ms usage is deprecated, please use HMS, HM or MS instead. Deprecated in version '1.5.6'.

data_pictures

#> # A tibble: 955 × 10
#>                                                        image_path
#>                                                             <chr>
#> 1  \\2015-06-04 08-36-32\\\\B00000001_21I6LW_20150603_084228E.JPG
#> 2                           B00000000_21I6LW_20150603_084228E.JPG
#> 3  \\2015-06-04 08-36-32\\\\B00000002_21I6LW_20150603_084236E.JPG
#> 4                           B00000003_21I6LW_20150603_084313E.JPG
#> 5                           B00000004_21I6LW_20150603_084349E.JPG
#> 6                           B00000005_21I6LW_20150603_084429E.JPG
#> 7                           B00000006_21I6LW_20150603_084509E.JPG
#> 8                           B00000007_21I6LW_20150603_084624E.JPG
#> 9                           B00000008_21I6LW_20150603_084657E.JPG
#> 10                          B00000009_21I6LW_20150603_084739E.JPG
#> # ... with 945 more rows, and 9 more variables: image_time <dttm>,
#> #   participant_id <chr>, `01a` <lgl>, `01b` <lgl>, `01c` <lgl>, `02b` <lgl>,
#> #   `02a` <lgl>, `03a` <lgl>, `01.` <lgl>

attr(data_pictures, "dico")

#>   Code            Meaning Group
#> 1  01a biomasscookingunit     1
#> 2  01b           lpgstove     1
#> 3  01c   othercookingunit     1
#> 4  02b    presencekitchen     1
#> 5  02a    foodpreparation     1
#> 6  03a             eating     1
#> 7  01.            cooking    99